Move user agent context functions to `useragent` package and deprecate old methods
